+++ api/avt/change.php
@@ -17,7 +17,7 @@
 
     checkToken();
 
-    if ($config["editinfo"] === false)
+    if ($config["editInfo"] === false)
         stop(21, "Thay đổi thông tin đã bị tắt!", 403);
     
     if (!isset($_FILES["file"]))